# 1132 improve api benchmarks