--- a/benchmarks/bench.html
+++ b/benchmarks/bench.html
@@ -6,28 +6,31 @@
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css">
     <title>User Metrics</title>
 </head>
+
 <body>
     <h1>Benchmarks</h1>
+    <p class="h6">Date: 2024-01-29 22:51:31</p>
+    <br>
 
 
-    GET /api/alive/
-    <table class="table">
+    <p  class="text-success"><strong>GET /api/alive/</strong></p>
+    <table class="table table-secondary">
         <tr>
             <th scope="col">
                 User
             </th>
             <th scope="col">
-                Time
+                Time(seconds)
             </th>
             <th scope="col">
-                Size
+                Size(byte)
             </th>
         </tr>
 
         <tr>
             <th>user-vanilla@example.org</th>
             <td>
-                0.0008738999999999998
+                0.000781
             </td>
             <td>
                 34.0
@@ -37,7 +40,7 @@
         <tr>
             <th>user-power@example.org</th>
             <td>
-                0.0008205
+                0.000755
             </td>
             <td>
                 34.0
@@ -47,24 +50,24 @@
 
     </table>
 
-    GET /api/v1/project/
-    <table class="table">
+    <p  class="text-success"><strong>GET /api/v1/project/</strong></p>
+    <table class="table table-secondary">
         <tr>
             <th scope="col">
                 User
             </th>
             <th scope="col">
-                Time
+                Time(seconds)
             </th>
             <th scope="col">
-                Size
+                Size(byte)
             </th>
         </tr>
 
         <tr>
             <th>user-vanilla@example.org</th>
             <td>
-                0.0046037000000000005
+                0.004588
             </td>
             <td>
                 96.0
@@ -74,7 +77,7 @@
         <tr>
             <th>user-power@example.org</th>
             <td>
-                0.004495600000000001
+                0.004412
             </td>
             <td>
                 96.0
@@ -84,24 +87,24 @@
 
     </table>
 
-    GET /api/v1/dataset/
-    <table class="table">
+    <p  class="text-success"><strong>GET /api/v1/dataset/</strong></p>
+    <table class="table table-secondary">
         <tr>
             <th scope="col">
                 User
             </th>
             <th scope="col">
-                Time
+                Time(seconds)
             </th>
             <th scope="col">
-                Size
+                Size(byte)
             </th>
         </tr>
 
         <tr>
             <th>user-vanilla@example.org</th>
             <td>
-                0.006133100000000001
+                0.006217
             </td>
             <td>
                 568.0
@@ -111,7 +114,7 @@
         <tr>
             <th>user-power@example.org</th>
             <td>
-                0.0791337
+                0.080017
             </td>
             <td>
                 141883.0