--- a/benchmarks/serve.sh
+++ b/benchmarks/serve.sh
@@ -6,13 +6,10 @@ WORKER_CLASS=uvicorn.workers.UvicornWorker
 
 fractalctl set-db
 
+python ../scripts/populate_db/populate_script_v2.py
+
 gunicorn "fractal_server.main:app" \
     --bind=$BIND \
     --workers=$N_WORKERS \
     --worker-class=$WORKER_CLASS \
     --daemon
-
-# let gunicorn takes its time
-sleep 3
-
-python populate_db.py