--- a/pyproject.toml
+++ b/pyproject.toml
@@ -59,7 +59,6 @@ pytest-subprocess = "^1.4"
 pytest-docker = {version = "^2.0"}
 pytest-pretty = "^1.0.1"
 pyyaml="^6.0.1"
-locust = "^2.20.0"
 a2wsgi = "^1.10.0"
 
 [tool.poetry.group.docs.dependencies]