@tcompa 2024-01-31
Example of new diff template
```
## Benchmark comparison
{% for (path_main, users_main), (path_curr, users_curr) in zip %}
### {{method}} {{ path_main }}

| User | Time(ms) Main | Time(ms) Current | Ratio Main/Current | Size(Kb) Main | Size(Kb) Current |
| -- | -- | -- | -- | -- | -- |
{% for i in range(users_main|length) %}| {{users_main[i].get("username")}} | {{users_main[i].get('time')}} | {{users_curr[i].get('time')}} | {{ users_main[i].get('time') / users_curr[i].get('time') }} | {{users_main[i].get('size')}} |{{users_curr[i].get('size')}} |
{%endfor%}

{% endfor %}
```

Which produces the following output (with fake data):

---

## Benchmark comparison

### GET /api/alive/

| User | Time(ms) Main | Time(ms) Current | Ratio Main/Current | Size(Kb) Main | Size(Kb) Current |
| -- | -- | -- | -- | -- | -- |
| user-vanilla | 0.8 | 0.8 | 1.0 | 0.0 |0.0 |
| user-power | 0.8 | 0.8 | 1.0 | 0.0 |0.0 |




### GET /api/v1/project/

| User | Time(ms) Main | Time(ms) Current | Ratio Main/Current | Size(Kb) Main | Size(Kb) Current |
| -- | -- | -- | -- | -- | -- |
| user-vanilla | 4.8 | 4.8 | 1.0 | 0.1 |0.1 |
| user-power | 4.7 | 4.7 | 1.0 | 0.1 |0.1 |




### GET /api/v1/dataset/

| User | Time(ms) Main | Time(ms) Current | Ratio Main/Current | Size(Kb) Main | Size(Kb) Current |
| -- | -- | -- | -- | -- | -- |
| user-vanilla | 7.0 | 7.0 | 1.0 | 0.6 |0.6 |
| user-power | 90.6 | 90.6 | 1.0 | 144.3 |144.3 |




### GET /api/v1/job/

| User | Time(ms) Main | Time(ms) Current | Ratio Main/Current | Size(Kb) Main | Size(Kb) Current |
| -- | -- | -- | -- | -- | -- |
| user-vanilla | 5.9 | 5.9 | 1.0 | 3.3 |3.3 |
| user-power | 119.4 | 119.4 | 1.0 | 504.4 |504.4 |